+++ cmd/nibid/cmd/root.go
@@ -139,13 +139,13 @@ func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {
 	cfg.Seal()
 
 	rootCmd.AddCommand(
-		genutilcli.InitCmd(app.ModuleBasics, app.DefaultNodeHome),
+		InitCmd(app.ModuleBasics, app.DefaultNodeHome),
 		genutilcli.CollectGenTxsCmd(banktypes.GenesisBalancesIterator{}, app.DefaultNodeHome),
 		genutilcli.MigrateGenesisCmd(),
 		genutilcli.GenTxCmd(app.ModuleBasics, encodingConfig.TxConfig, banktypes.GenesisBalancesIterator{}, app.DefaultNodeHome),
 		genutilcli.ValidateGenesisCmd(app.ModuleBasics),
 		AddGenesisAccountCmd(app.DefaultNodeHome),
-		vpoolcli.AddVPoolGenesisCmd(app.DefaultNodeHome),
+		vpoolcli.AddVpoolGenesisCmd(app.DefaultNodeHome),
 		pricefeedcli.AddWhitelistGenesisOracle(app.DefaultNodeHome),
 		pricefeedcli.AddPriceFeedParamPairs(app.DefaultNodeHome),
 		tmcli.NewCompletionCmd(rootCmd, true),