+++ x/oracle/keeper/querier.go
@@ -3,12 +3,11 @@ package keeper
 import (
 	"context"
 
-	"github.com/NibiruChain/nibiru/collections"
-
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/NibiruChain/collections"
 
 	"github.com/NibiruChain/nibiru/x/oracle/types"
 )
@@ -77,7 +76,7 @@ func (q querier) Actives(c context.Context, _ *types.QueryActivesRequest) (*type
 // VoteTargets queries the voting target list on current vote period
 func (q querier) VoteTargets(c context.Context, _ *types.QueryVoteTargetsRequest) (*types.QueryVoteTargetsResponse, error) {
 	ctx := sdk.UnwrapSDKContext(c)
-	return &types.QueryVoteTargetsResponse{VoteTargets: q.GetVoteTargets(ctx)}, nil
+	return &types.QueryVoteTargetsResponse{VoteTargets: q.GetWhitelistedPairs(ctx)}, nil
 }
 
 // FeederDelegation queries the account address that the validator operator delegated oracle vote rights to