+++ x/stablecoin/types/expected_keepers.go
@@ -46,7 +46,7 @@ type PricefeedKeeper interface {
 }
 
 type DexKeeper interface {
-	GetFromPair(ctx sdk.Context, denomA string, denomB string,
-	) (poolId uint64, err error)
+	FetchPoolFromPair(ctx sdk.Context, denomA string, denomB string,
+	) (pool dextypes.Pool, err error)
 	FetchPool(ctx sdk.Context, poolId uint64) (pool dextypes.Pool, err error)
 }