+++ x/vpool/keeper/pool_state.go
@@ -1,11 +1,13 @@
 package keeper
 
 import (
+	"fmt"
+	"math"
 	"time"
 
-	"github.com/NibiruChain/nibiru/collections"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/NibiruChain/collections"
 
 	"github.com/NibiruChain/nibiru/x/common"
 	"github.com/NibiruChain/nibiru/x/vpool/types"
@@ -15,35 +17,106 @@ import (
 func (k Keeper) CreatePool(
 	ctx sdk.Context,
 	pair common.AssetPair,
-	tradeLimitRatio sdk.Dec, // integer with 6 decimals, 1_000_000 means 1.0
 	quoteAssetReserve sdk.Dec,
 	baseAssetReserve sdk.Dec,
-	fluctuationLimitRatio sdk.Dec,
-	maxOracleSpreadRatio sdk.Dec,
-	maintenanceMarginRatio sdk.Dec,
-	maxLeverage sdk.Dec,
+	config types.VpoolConfig,
 ) {
-	k.Pools.Insert(ctx, pair, types.VPool{
-		Pair:                   pair,
-		BaseAssetReserve:       baseAssetReserve,
-		QuoteAssetReserve:      quoteAssetReserve,
-		TradeLimitRatio:        tradeLimitRatio,
-		FluctuationLimitRatio:  fluctuationLimitRatio,
-		MaxOracleSpreadRatio:   maxOracleSpreadRatio,
-		MaintenanceMarginRatio: maintenanceMarginRatio,
-		MaxLeverage:            maxLeverage,
-	})
+	vpool := types.Vpool{
+		Pair:              pair,
+		BaseAssetReserve:  baseAssetReserve,
+		QuoteAssetReserve: quoteAssetReserve,
+		Config:            config,
+	}
+	k.Pools.Insert(ctx, pair, vpool)
 
 	k.ReserveSnapshots.Insert(
 		ctx,
 		collections.Join(pair, ctx.BlockTime()),
-		types.NewReserveSnapshot(
-			pair,
-			baseAssetReserve,
-			quoteAssetReserve,
-			ctx.BlockTime(),
-		),
+		vpool.ToSnapshot(ctx),
 	)
+}
+
+func (k Keeper) EditPoolConfig(
+	ctx sdk.Context,
+	pair common.AssetPair,
+	config types.VpoolConfig,
+) error {
+	// Grab current pool from state
+	vpool, err := k.Pools.Get(ctx, pair)
+	if err != nil {
+		return err
+	}
+
+	newVpool := types.Vpool{
+		Pair:              vpool.Pair,
+		BaseAssetReserve:  vpool.BaseAssetReserve,
+		QuoteAssetReserve: vpool.QuoteAssetReserve,
+		Config:            config, // main change is here
+	}
+	if err := newVpool.Validate(); err != nil {
+		return err
+	}
+
+	err = k.updatePool(
+		ctx,
+		newVpool,
+		/*skipFluctuationLimitCheck*/ true)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (k Keeper) EditSwapInvariant(
+	ctx sdk.Context,
+	swapInvariantMap types.EditSwapInvariantsProposal_SwapInvariantMultiple,
+) error {
+	if err := swapInvariantMap.Validate(); err != nil {
+		return err
+	}
+
+	// Grab current pool from state
+	vpool, err := k.Pools.Get(ctx, common.MustNewAssetPair(swapInvariantMap.Pair))
+	if err != nil {
+		return err
+	}
+
+	// price = y / x
+	// k = x * y
+	// newK = (cx) * (cy) = c^2 xy = c^2 k
+	// newPrice = (c y) / (c x) = y / x = price
+	swapInvariant := vpool.BaseAssetReserve.Mul(vpool.QuoteAssetReserve)
+	newSwapInvariant := swapInvariant.Mul(swapInvariantMap.Multiplier)
+
+	// Change the swap invariant while holding price constant.
+	// Multiplying by the same factor to both of the reserves won't affect price.
+	cSquared := newSwapInvariant.Quo(swapInvariant).MustFloat64()
+	cAsFloat := math.Sqrt(cSquared)
+	c, err := sdk.NewDecFromStr(fmt.Sprintf("%f", cAsFloat))
+	if err != nil {
+		return err
+	}
+	newBaseAmount := c.Mul(vpool.BaseAssetReserve)
+	newQuoteAmount := c.Mul(vpool.QuoteAssetReserve)
+
+	newVpool := types.Vpool{
+		Pair:              vpool.Pair,
+		BaseAssetReserve:  newBaseAmount,
+		QuoteAssetReserve: newQuoteAmount,
+		Config:            vpool.Config,
+	}
+	if err := newVpool.Validate(); err != nil {
+		return err
+	}
+
+	err = k.updatePool(
+		ctx,
+		newVpool,
+		/*skipFluctuationLimitCheck*/ true)
+	if err != nil {
+		return err
+	}
+	return nil
 }
 
 /*
@@ -59,7 +132,7 @@ ret:
 */
 func (k Keeper) updatePool(
 	ctx sdk.Context,
-	updatedPool types.VPool,
+	updatedPool types.Vpool,
 	skipFluctuationCheck bool,
 ) (err error) {
 	if !skipFluctuationCheck {
@@ -83,7 +156,7 @@ func (k Keeper) ExistsPool(ctx sdk.Context, pair common.AssetPair) bool {
 // An error is returned if the pool does not exist.
 // No error is returned if the prices don't exist, however.
 func (k Keeper) GetPoolPrices(
-	ctx sdk.Context, pool types.VPool,
+	ctx sdk.Context, pool types.Vpool,
 ) (types.PoolPrices, error) {
 	if err := pool.Pair.Validate(); err != nil {
 		return types.PoolPrices{}, err