+++ x/vpool/keeper/query_server.go
@@ -3,7 +3,7 @@ package keeper
 import (
 	"context"
 
-	"github.com/NibiruChain/nibiru/collections"
+	"github.com/NibiruChain/collections"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"google.golang.org/grpc/codes"
@@ -59,7 +59,7 @@ func (q queryServer) AllPools(
 
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
-	var pools []types.VPool
+	var pools []types.Vpool
 	var pricesForPools []types.PoolPrices
 	for _, pool := range q.k.Pools.Iterate(ctx, collections.Range[common.AssetPair]{}).Values() {
 		poolPrices, err := q.k.GetPoolPrices(ctx, pool)