+++ x/vpool/types/codec.go
@@ -17,6 +17,8 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 	)
 
 	registry.RegisterImplementations((*govtypes.Content)(nil), &CreatePoolProposal{})
+	registry.RegisterImplementations((*govtypes.Content)(nil), &EditPoolConfigProposal{})
+	registry.RegisterImplementations((*govtypes.Content)(nil), &EditSwapInvariantsProposal{})
 
 	// msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)
 }